cloudapp/requirements.txt
flask ~=3.0.3
aws_lambda_wsgi ~=0.0.6
azure_functions ~=1.19.0
python_dotenv ~=1.0.1